Add script for building project Some refactoring

+++ src/main/java/com/epam/action/TableAction.java
@@ -4,7 +4,7 @@ import com.epam.entity.Table;
 import com.epam.service.impservice.TableService;
 
 public class TableAction {
-    public int reserveTable(Table table) {
+    public long reserveTable(Table table) {
         TableService tableService = new TableService();
         return tableService.reserve(table);
     }
